tests/unit/routes/cartRoutes.test.ts
// tests/unit/routes/cartRoutes.test.ts

import request from 'supertest';
import express from 'express';
import cartRoutes from '../../../src/routes/cartRoutes';

const app = express();
app.use(express.json()); // Enable JSON parsing for requests
app.use('/api/cart', cartRoutes); // Mount the cart routes

describe('Cart Routes', () => {
  it('should get the cart', async () => {
    const response = await request(app).get('/api/cart');
    expect(response.status).toBe(200);
    // Add more assertions based on expected response structure
  });

  it('should add an item to the cart', async () => {
    const response = await request(app)
      .post('/api/cart/add')
      .send({ productId: '12345', quantity: 1 });
    expect(response.status).toBe(201);
    // Add more assertions based on expected response structure
  });

  // Add more tests for update, delete, sync, and merge functionalities
});
